requirements.txt
torchvision==0.3.0
tifffile==2019.7.26.2
setproctitle==1.1.10
numpy==1.17.3
opencv_python==3.4.1.15
torch==1.1.0
Pillow==6.2.1
